requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.1 